requirements.txt
aiofiles
git+https://github.com/agronholm/anyio
aiohttp
bs4==0.0.1
datetime
git+https://github.com/Joelkb/cinemagoer
pymongo[srv]==3.12.3
motor==2.5.1
marshmallow==3.14.1
umongo==3.0.1
psutil==5.9.4
pyrofork
python-dotenv
pytz
requests
tgcrypto